Core: added cast to long type.
This related #129 issue on GitHub.

+++ source/lexbor/core/conv.h
@@ -42,7 +42,7 @@ lexbor_conv_double_to_long(double number)
         return -LONG_MAX;
     }
 
-    return number;
+    return (long) number;
 }
